Move CTransfer_FlowTraction and CTransfer_StructuralDisplacements
