+++ SU2_CFD/include/transfer/CTransfer.hpp
@@ -245,262 +245,3 @@ public:
 
 };
 
-/*!
- * \class CTransfer_FlowTraction
- * \brief Transfer flow tractions from a fluid zone into a structural zone
- * \author R. Sanchez
- * \version 6.2.0 "Falcon"
- */
-
-class CTransfer_FlowTraction : public CTransfer {
-
-protected:
-  bool consistent_interpolation;
-  
-  /*!
-   * \brief Sets the dimensional factor for pressure and the consistent_interpolation flag
-   * \param[in] flow_config - Definition of the fluid (donor) problem.
-   */
-  void Preprocess(CConfig *flow_config);
-
-public:
-
-  /*!
-   * \brief Constructor of the class.
-   */
-  CTransfer_FlowTraction(void);
-
-  /*!
-   * \overload
-   * \param[in] val_nVar - Number of variables that need to be transferred.
-   * \param[in] config - Definition of the particular problem.
-   */
-  CTransfer_FlowTraction(unsigned short val_nVar, unsigned short val_nConst, CConfig *config);
-
-  /*!
-   * \brief Destructor of the class.
-   */
-  virtual ~CTransfer_FlowTraction(void);
-
-  /*!
-   * \brief Retrieve some constants needed for the calculations.
-   * \param[in] donor_solution - Solution from the donor mesh.
-   * \param[in] target_solution - Solution from the target mesh.
-   * \param[in] donor_geometry - Geometry of the donor mesh.
-   * \param[in] target_geometry - Geometry of the target mesh.
-   * \param[in] donor_config - Definition of the problem at the donor mesh.
-   * \param[in] target_config - Definition of the problem at the target mesh.
-   */
-  void GetPhysical_Constants(CSolver *donor_solution, CSolver *target_solution,
-                 CGeometry *donor_geometry, CGeometry *target_geometry,
-                 CConfig *donor_config, CConfig *target_config);
-
-  /*!
-   * \brief Retrieve the variable that will be sent from donor mesh to target mesh.
-   * \param[in] donor_solution - Solution from the donor mesh.
-   * \param[in] donor_geometry - Geometry of the donor mesh.
-   * \param[in] donor_config - Definition of the problem at the donor mesh.
-   * \param[in] Marker_Donor - Index of the donor marker.
-   * \param[in] Vertex_Donor - Index of the donor vertex.
-   */
-  void GetDonor_Variable(CSolver *flow_solution, CGeometry *flow_geometry, CConfig *flow_config,
-               unsigned long Marker_Flow, unsigned long Vertex_Flow, unsigned long Point_Flow);
-
-  /*!
-   * \brief Set the variable that has been received from the target mesh into the target mesh.
-   * \param[in] target_solution - Solution from the target mesh.
-   * \param[in] target_geometry - Geometry of the target mesh.
-   * \param[in] target_config - Definition of the problem at the target mesh.
-   * \param[in] Marker_Target - Index of the target marker.
-   * \param[in] Vertex_Target - Index of the target vertex.
-   * \param[in] Point_Target - Index of the target point.
-   */
-  void SetTarget_Variable(CSolver *fea_solution, CGeometry *fea_geometry,
-              CConfig *fea_config, unsigned long Marker_Struct,
-              unsigned long Vertex_Struct, unsigned long Point_Struct);
-
-};
-
-/*!
- * \class CTransfer_StructuralDisplacements
- * \brief Transfer structural displacements from a structural zone into a fluid zone
- * \author R. Sanchez
- * \version 6.2.0 "Falcon"
- */
-
-class CTransfer_StructuralDisplacements : public CTransfer {
-
-protected:
-
-public:
-
-  /*!
-   * \brief Constructor of the class.
-   */
-  CTransfer_StructuralDisplacements(void);
-
-  /*!
-   * \overload
-   * \param[in] val_nVar - Number of variables that need to be transferred.
-   * \param[in] config - Definition of the particular problem.
-   */
-  CTransfer_StructuralDisplacements(unsigned short val_nVar, unsigned short val_nConst, CConfig *config);
-
-  /*!
-   * \brief Destructor of the class.
-   */
-  virtual ~CTransfer_StructuralDisplacements(void);
-
-  /*!
-   * \brief Retrieve some constants needed for the calculations.
-   * \param[in] donor_solution - Solution from the donor mesh.
-   * \param[in] target_solution - Solution from the target mesh.
-   * \param[in] donor_geometry - Geometry of the donor mesh.
-   * \param[in] target_geometry - Geometry of the target mesh.
-   * \param[in] donor_config - Definition of the problem at the donor mesh.
-   * \param[in] target_config - Definition of the problem at the target mesh.
-   */
-  void GetPhysical_Constants(CSolver *donor_solution, CSolver *target_solution,
-                 CGeometry *donor_geometry, CGeometry *target_geometry,
-                 CConfig *donor_config, CConfig *target_config);
-
-  /*!
-   * \brief Retrieve the variable that will be sent from donor mesh to target mesh.
-   * \param[in] donor_solution - Solution from the donor mesh.
-   * \param[in] donor_geometry - Geometry of the donor mesh.
-   * \param[in] donor_config - Definition of the problem at the donor mesh.
-   * \param[in] Marker_Donor - Index of the donor marker.
-   * \param[in] Vertex_Donor - Index of the donor vertex.
-   */
-  void GetDonor_Variable(CSolver *struct_solution, CGeometry *struct_geometry, CConfig *struct_config,
-               unsigned long Marker_Struct, unsigned long Vertex_Struct, unsigned long Point_Struct);
-
-  /*!
-   * \brief Set the variable that has been received from the target mesh into the target mesh.
-   * \param[in] target_solution - Solution from the target mesh.
-   * \param[in] target_geometry - Geometry of the target mesh.
-   * \param[in] target_config - Definition of the problem at the target mesh.
-   * \param[in] Marker_Target - Index of the target marker.
-   * \param[in] Vertex_Target - Index of the target vertex.
-   * \param[in] Point_Target - Index of the target point.
-   */
-  void SetTarget_Variable(CSolver *flow_solution, CGeometry *flow_geometry,
-              CConfig *flow_config, unsigned long Marker_Flow,
-              unsigned long Vertex_Flow, unsigned long Point_Flow);
-
-};
-
-/*!
- * \class CTransfer_FlowTraction_DiscAdj
- * \brief Transfer flow tractions from a fluid zone into a structural zone in a discrete adjoint simulation
- * \author R. Sanchez
- * \version 6.2.0 "Falcon"
- */
-
-class CTransfer_FlowTraction_DiscAdj : public CTransfer_FlowTraction {
-
-protected:
-
-public:
-
-  /*!
-   * \brief Constructor of the class.
-   */
-  CTransfer_FlowTraction_DiscAdj(void);
-
-  /*!
-   * \overload
-   * \param[in] val_nVar - Number of variables that need to be transferred.
-   * \param[in] config - Definition of the particular problem.
-   */
-  CTransfer_FlowTraction_DiscAdj(unsigned short val_nVar, unsigned short val_nConst, CConfig *config);
-
-  /*!
-   * \brief Destructor of the class.
-   */
-  virtual ~CTransfer_FlowTraction_DiscAdj(void);
-
-  /*!
-   * \brief Retrieve some constants needed for the calculations.
-   * \param[in] donor_solution - Solution from the donor mesh.
-   * \param[in] target_solution - Solution from the target mesh.
-   * \param[in] donor_geometry - Geometry of the donor mesh.
-   * \param[in] target_geometry - Geometry of the target mesh.
-   * \param[in] donor_config - Definition of the problem at the donor mesh.
-   * \param[in] target_config - Definition of the problem at the target mesh.
-   */
-  void GetPhysical_Constants(CSolver *donor_solution, CSolver *target_solution,
-                 CGeometry *donor_geometry, CGeometry *target_geometry,
-                 CConfig *donor_config, CConfig *target_config);
-
-};
-
-/*!
- * \class CTransfer_StructuralDisplacements_DiscAdj
- * \brief Transfer structural displacements from a structural zone into a fluid zone in a discrete adjoint simulation
- * \author R. Sanchez
- * \version 6.2.0 "Falcon"
- */
-
-class CTransfer_StructuralDisplacements_DiscAdj : public CTransfer {
-
-protected:
-
-public:
-
-  /*!
-   * \brief Constructor of the class.
-   */
-  CTransfer_StructuralDisplacements_DiscAdj(void);
-
-  /*!
-   * \overload
-   * \param[in] val_nVar - Number of variables that need to be transferred.
-   * \param[in] config - Definition of the particular problem.
-   */
-  CTransfer_StructuralDisplacements_DiscAdj(unsigned short val_nVar, unsigned short val_nConst, CConfig *config);
-
-  /*!
-   * \brief Destructor of the class.
-   */
-  virtual ~CTransfer_StructuralDisplacements_DiscAdj(void);
-
-  /*!
-   * \brief Retrieve some constants needed for the calculations.
-   * \param[in] donor_solution - Solution from the donor mesh.
-   * \param[in] target_solution - Solution from the target mesh.
-   * \param[in] donor_geometry - Geometry of the donor mesh.
-   * \param[in] target_geometry - Geometry of the target mesh.
-   * \param[in] donor_config - Definition of the problem at the donor mesh.
-   * \param[in] target_config - Definition of the problem at the target mesh.
-   */
-  void GetPhysical_Constants(CSolver *donor_solution, CSolver *target_solution,
-                 CGeometry *donor_geometry, CGeometry *target_geometry,
-                 CConfig *donor_config, CConfig *target_config);
-
-  /*!
-   * \brief Retrieve the variable that will be sent from donor mesh to target mesh.
-   * \param[in] donor_solution - Solution from the donor mesh.
-   * \param[in] donor_geometry - Geometry of the donor mesh.
-   * \param[in] donor_config - Definition of the problem at the donor mesh.
-   * \param[in] Marker_Donor - Index of the donor marker.
-   * \param[in] Vertex_Donor - Index of the donor vertex.
-   */
-  void GetDonor_Variable(CSolver *struct_solution, CGeometry *struct_geometry, CConfig *struct_config,
-               unsigned long Marker_Struct, unsigned long Vertex_Struct, unsigned long Point_Struct);
-
-  /*!
-   * \brief Set the variable that has been received from the target mesh into the target mesh.
-   * \param[in] target_solution - Solution from the target mesh.
-   * \param[in] target_geometry - Geometry of the target mesh.
-   * \param[in] target_config - Definition of the problem at the target mesh.
-   * \param[in] Marker_Target - Index of the target marker.
-   * \param[in] Vertex_Target - Index of the target vertex.
-   * \param[in] Point_Target - Index of the target point.
-   */
-  void SetTarget_Variable(CSolver *flow_solution, CGeometry *flow_geometry,
-              CConfig *flow_config, unsigned long Marker_Flow,
-              unsigned long Vertex_Flow, unsigned long Point_Flow);
-
-};
-
